ISSUE #1659: Introduced ActivityIdRepository
