+++ src/Domain/Activity/DbalActivityRepository.php
@@ -6,7 +6,6 @@ use App\Domain\Activity\Route\RouteGeography;
 use App\Domain\Activity\SportType\SportType;
 use App\Domain\Activity\SportType\SportTypes;
 use App\Domain\Gear\GearId;
-use App\Domain\Gear\GearIds;
 use App\Infrastructure\Exception\EntityNotFound;
 use App\Infrastructure\Serialization\Json;
 use App\Infrastructure\ValueObject\Geography\Coordinate;
@@ -15,7 +14,6 @@ use App\Infrastructure\ValueObject\Geography\Longitude;
 use App\Infrastructure\ValueObject\Measurement\Length\Meter;
 use App\Infrastructure\ValueObject\Measurement\Velocity\KmPerHour;
 use App\Infrastructure\ValueObject\Time\SerializableDateTime;
-use App\Infrastructure\ValueObject\Time\Years;
 use Doctrine\DBAL\ArrayParameterType;
 use Doctrine\DBAL\Connection;
 
@@ -60,38 +58,6 @@ final class DbalActivityRepository implements ActivityRepository
             name: $result['name'],
             startDateTime: SerializableDateTime::fromString($result['startDateTime']),
         );
-    }
-
-    public function findLongestActivityFor(Years $years): Activity
-    {
-        if (!$result = $this->connection->executeQuery(
-            <<<SQL
-                SELECT *
-                FROM Activity
-                WHERE strftime('%Y',startDateTime) IN (:years)
-                ORDER BY movingTimeInSeconds DESC
-                LIMIT 1
-            SQL,
-            [
-                'years' => array_map(strval(...), $years->toArray()),
-            ],
-            [
-                'years' => ArrayParameterType::STRING,
-            ]
-        )->fetchAssociative()) {
-            throw new EntityNotFound('Could not determine longest activity');
-        }
-
-        return $this->hydrate($result);
-    }
-
-    public function count(): int
-    {
-        $queryBuilder = $this->connection->createQueryBuilder();
-        $queryBuilder->select('COUNT(*)')
-            ->from('Activity');
-
-        return (int) $queryBuilder->executeQuery()->fetchOne();
     }
 
     public function findAll(?int $limit = null): Activities
@@ -165,69 +131,6 @@ final class DbalActivityRepository implements ActivityRepository
         return Activities::fromArray(array_map(
             $this->hydrate(...),
             $queryBuilder->executeQuery()->fetchAllAssociative()
-        ));
-    }
-
-    public function hasForSportTypes(SportTypes $sportTypes): bool
-    {
-        $queryBuilder = $this->connection->createQueryBuilder();
-        $queryBuilder->select('COUNT(*)')
-            ->from('Activity')
-            ->andWhere('sportType IN (:sportTypes)')
-            ->setParameter(
-                key: 'sportTypes',
-                value: array_map(fn (SportType $sportType) => $sportType->value, $sportTypes->toArray()),
-                type: ArrayParameterType::STRING
-            );
-
-        return (bool) $queryBuilder->executeQuery()->fetchOne();
-    }
-
-    public function findActivityIds(): ActivityIds
-    {
-        $queryBuilder = $this->connection->createQueryBuilder();
-        $queryBuilder->select('activityId')
-            ->from('Activity')
-            ->orderBy('startDateTime', 'DESC');
-
-        return ActivityIds::fromArray(array_map(
-            ActivityId::fromString(...),
-            $queryBuilder->executeQuery()->fetchFirstColumn(),
-        ));
-    }
-
-    public function findUniqueStravaGearIds(?ActivityIds $restrictToActivityIds): GearIds
-    {
-        $queryBuilder = $this->connection->createQueryBuilder();
-        $queryBuilder->select('DISTINCT JSON_EXTRACT(data, "$.gear_id") as stravaGearId')
-            ->from('Activity')
-            ->andWhere('stravaGearId IS NOT NULL');
-
-        if ($restrictToActivityIds && !$restrictToActivityIds->isEmpty()) {
-            $queryBuilder->andWhere('activityId IN (:activityIds)');
-            $queryBuilder->setParameter(
-                key: 'activityIds',
-                value: array_map(strval(...), $restrictToActivityIds->toArray()),
-                type: ArrayParameterType::STRING
-            );
-        }
-
-        return GearIds::fromArray(array_map(
-            GearId::fromUnprefixed(...),
-            $queryBuilder->executeQuery()->fetchFirstColumn(),
-        ));
-    }
-
-    public function findActivityIdsMarkedForDeletion(): ActivityIds
-    {
-        $queryBuilder = $this->connection->createQueryBuilder();
-        $queryBuilder->select('activityId')
-            ->from('Activity')
-            ->where('markedForDeletion = 1');
-
-        return ActivityIds::fromArray(array_map(
-            ActivityId::fromString(...),
-            $queryBuilder->executeQuery()->fetchFirstColumn(),
         ));
     }
