+++ src/Domain/Gear/ImportedGear/ImportedGearRepository.php
@@ -2,7 +2,9 @@
 
 namespace App\Domain\Gear\ImportedGear;
 
+use App\Domain\Activity\ActivityIds;
 use App\Domain\Gear\GearId;
+use App\Domain\Gear\GearIds;
 use App\Domain\Gear\Gears;
 
 interface ImportedGearRepository
@@ -14,4 +16,6 @@ interface ImportedGearRepository
     public function findAllUsed(): Gears;
 
     public function find(GearId $gearId): ImportedGear;
+
+    public function findUniqueStravaGearIds(?ActivityIds $restrictToActivityIds): GearIds;
 }
