+++ tests/Domain/Activity/DbalActivityRepositoryTest.php
@@ -4,7 +4,6 @@ namespace App\Tests\Domain\Activity;
 
 use App\Domain\Activity\Activity;
 use App\Domain\Activity\ActivityId;
-use App\Domain\Activity\ActivityIds;
 use App\Domain\Activity\ActivityRepository;
 use App\Domain\Activity\ActivitySummary;
 use App\Domain\Activity\ActivityType;
@@ -14,11 +13,8 @@ use App\Domain\Activity\DbalActivityRepository;
 use App\Domain\Activity\DbalActivityWithRawDataRepository;
 use App\Domain\Activity\SportType\SportType;
 use App\Domain\Activity\SportType\SportTypes;
-use App\Domain\Gear\GearId;
 use App\Infrastructure\Exception\EntityNotFound;
 use App\Infrastructure\ValueObject\Time\SerializableDateTime;
-use App\Infrastructure\ValueObject\Time\Year;
-use App\Infrastructure\ValueObject\Time\Years;
 use App\Tests\ContainerTestCase;
 use Spatie\Snapshots\MatchesSnapshots;
 
@@ -185,142 +181,6 @@ class DbalActivityRepositoryTest extends ContainerTestCase
             $this->activityRepository->findBySportTypes(SportTypes::fromArray(
                 [SportType::RUN, SportType::MOUNTAIN_BIKE_RIDE]
             ))->map(fn (Activity $activity): ActivityId => $activity->getId())
-        );
-    }
-
-    public function testFindLongestActivityFor(): void
-    {
-        $longestActivity = ActivityBuilder::fromDefaults()
-            ->withActivityId(ActivityId::fromUnprefixed('0'))
-            ->withMovingTimeInSeconds(10000)
-            ->withStartDateTime(SerializableDateTime::fromString('2024-03-01 00:00:00'))
-            ->build();
-
-        $this->getContainer()->get(ActivityWithRawDataRepository::class)->add(ActivityWithRawData::fromState(
-            $longestActivity,
-            []
-        ));
-
-        $this->getContainer()->get(ActivityWithRawDataRepository::class)->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('1'))
-                ->withMovingTimeInSeconds(20000)
-                ->withStartDateTime(SerializableDateTime::fromString('2025-01-01 00:00:00'))
-                ->build(),
-            []
-        ));
-        $this->getContainer()->get(ActivityWithRawDataRepository::class)->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('2'))
-                ->withGearId(GearId::fromUnprefixed('3'))
-                ->withStartDateTime(SerializableDateTime::fromString('2023-01-01 00:00:00'))
-                ->build(),
-            []
-        ));
-        $this->getContainer()->get(ActivityWithRawDataRepository::class)->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('3'))
-                ->withGearId(GearId::fromUnprefixed('2'))
-                ->withStartDateTime(SerializableDateTime::fromString('2024-01-01 00:00:00'))
-                ->build(),
-            []
-        ));
-        $this->getContainer()->get(ActivityWithRawDataRepository::class)->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('4'))
-                ->withGearId(GearId::fromUnprefixed('5'))
-                ->withStartDateTime(SerializableDateTime::fromString('2024-01-03 00:00:00'))
-                ->build(),
-            []
-        ));
-        $this->getContainer()->get(ActivityWithRawDataRepository::class)->add(ActivityWithRawData::fromState(
-            ActivityBuilder::fromDefaults()
-                ->withActivityId(ActivityId::fromUnprefixed('8'))
-                ->withGearId(GearId::fromUnprefixed('5'))
-                ->withStartDateTime(SerializableDateTime::fromString('2024-01-03 00:00:00'))
-                ->build(),
-            []
-        ));
-
-        $this->assertEquals(
-            $longestActivity,
-            $this->activityRepository->findLongestActivityFor(Years::fromArray([Year::fromInt(2024)])),
-        );
-    }
-
-    public function testFindLongestActivityForYearItShouldThrow(): void
-    {
-        $this->expectExceptionObject(new EntityNotFound('Could not determine longest activity'));
-        $this->activityRepository->findLongestActivityFor(Years::fromArray([Year::fromInt(2024)]));
-    }
-
-    public function testCount(): void
-    {
-        $activityOne = ActivityBuilder::fromDefaults()
-            ->withActivityId(ActivityId::fromUnprefixed(1))
-            ->withStartDateTime(SerializableDateTime::fromString('2023-10-10 14:00:34'))
-            ->build();
-        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
-            $activityOne,
-            ['raw' => 'data']
-        ));
-        $activityTwo = ActivityBuilder::fromDefaults()
-            ->withActivityId(ActivityId::fromUnprefixed(2))
-            ->withStartDateTime(SerializableDateTime::fromString('2023-10-10 13:00:34'))
-            ->build();
-        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
-            $activityTwo,
-            ['raw' => 'data']
-        ));
-        $activityThree = ActivityBuilder::fromDefaults()
-            ->withActivityId(ActivityId::fromUnprefixed(3))
-            ->withStartDateTime(SerializableDateTime::fromString('2023-10-09 14:00:34'))
-            ->build();
-        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
-            $activityThree,
-            ['raw' => 'data']
-        ));
-
-        $this->assertEquals(
-            3,
-            $this->activityRepository->count()
-        );
-    }
-
-    public function testFindActivityIds(): void
-    {
-        $activityOne = ActivityBuilder::fromDefaults()
-            ->withActivityId(ActivityId::fromUnprefixed(1))
-            ->withStartDateTime(SerializableDateTime::fromString('2023-10-10 14:00:34'))
-            ->build();
-        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
-            $activityOne,
-            ['raw' => 'data']
-        ));
-        $activityTwo = ActivityBuilder::fromDefaults()
-            ->withActivityId(ActivityId::fromUnprefixed(2))
-            ->withStartDateTime(SerializableDateTime::fromString('2023-10-10 13:00:34'))
-            ->build();
-        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
-            $activityTwo,
-            ['raw' => 'data']
-        ));
-        $activityThree = ActivityBuilder::fromDefaults()
-            ->withActivityId(ActivityId::fromUnprefixed(3))
-            ->withStartDateTime(SerializableDateTime::fromString('2023-10-09 14:00:34'))
-            ->build();
-        $this->activityWithRawDataRepository->add(ActivityWithRawData::fromState(
-            $activityThree,
-            ['raw' => 'data']
-        ));
-
-        $this->assertEquals(
-            ActivityIds::fromArray([
-                ActivityId::fromUnprefixed(1),
-                ActivityId::fromUnprefixed(2),
-                ActivityId::fromUnprefixed(3),
-            ]),
-            $this->activityRepository->findActivityIds()
         );
     }
