+++ tests/Domain/Challenge/Consistency/ConsistencyChallengeCalculatorTest.php
@@ -2,7 +2,7 @@
 
 namespace App\Tests\Domain\Challenge\Consistency;
 
-use App\Domain\Activity\ActivityRepository;
+use App\Domain\Activity\ActivityIdRepository;
 use App\Domain\Activity\SportType\SportType;
 use App\Domain\Activity\SportType\SportTypes;
 use App\Domain\Calendar\Months;
@@ -136,7 +136,7 @@ class ConsistencyChallengeCalculatorTest extends ContainerTestCase
         parent::setUp();
 
         $this->calculator = new ConsistencyChallengeCalculator(
-            $this->getContainer()->get(ActivityRepository::class),
+            $this->getContainer()->get(ActivityIdRepository::class),
             $this->getContainer()->get(QueryBus::class),
         );
     }
